add wayland warning

+++ assists/c.sh
@@ -2,6 +2,11 @@
 
 # assist: toggle compositing
 
+if [ -z "$XDG_SESSION_TYPE" = "wayland" ]; then
+    notify-send -a instantASSIST ' you on wayland you doofus'
+    exit
+fi
+
 if pgrep picom; then
     notify-send -a instantASSIST ' compositing disabled'
     pkill picom
